feat(outline): merge chpter:problem & chapter:analysis
Initial plan was to talk about the issues without saying anything
concrete, then in the analysis chapter demonstrate the impact of all
those issues on a concrete system. But now I realized that talking about
the problems without showing any trace or anything concrete will be
quite difficult. So we'll have a single chapter to cover it all.

It might be the biggest chapter in the dissertation.